requirements.txt
torch
torchvision
facenet-pytorch
flask
flask_sqlalchemy
opencv-python
pillow
numpy